Fix `ssh2` not being able to create new directories on Windows
There is a problem with the `ssh2` crate (both with scp and sftp) that
we can not create new directories, so instead we have to rely on pre-existing
directories if we want to create / upload files to the Windows guest. As a
temporary workaround, use `C:` as a temporary directory.